rename block2 -> block
